# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/register": {
      "post": {
        "operationId": "Auth_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "summary": "Đăng ký tài khoản hành khách hoặc tài xế",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "Auth_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "summary": "Đăng nhập bằng email + mật khẩu, trả về JWT",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "Auth_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Thông tin người dùng hiện tại",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Share Car API",
    "description": "API đặt xe ghép 5/7 chỗ — hành khách, tài xế, admin",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RegisterDto": {
        "type": "object",
        "properties": {
          "role": {
            "description": "Chỉ cho phép tự đăng ký PASSENGER hoặc DRIVER. ADMIN được tạo qua seed/CLI.",
            "enum": [
              "PASSENGER",
              "DRIVER"
            ],
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string",
            "description": "Số điện thoại Việt Nam, ví dụ 0912345678",
            "pattern": "^0\\d{9}$"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 72
          },
          "fullName": {
            "type": "string",
            "maxLength": 100
          },
          "licenseNumber": {
            "type": "string",
            "description": "Số giấy phép lái xe — bắt buộc khi role = DRIVER",
            "maxLength": 30
          }
        },
        "required": [
          "role",
          "email",
          "phone",
          "password",
          "fullName"
        ]
      },
      "Role": {
        "type": "string",
        "enum": [
          "PASSENGER",
          "DRIVER",
          "ADMIN"
        ]
      },
      "UserStatus": {
        "type": "string",
        "enum": [
          "ACTIVE",
          "LOCKED"
        ]
      },
      "DriverStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "APPROVED",
          "REJECTED"
        ]
      },
      "DriverProfileDto": {
        "type": "object",
        "properties": {
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/DriverStatus"
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "licenseNumber": {
            "type": "string"
          },
          "rejectReason": {
            "type": "string",
            "nullable": true
          },
          "reviewedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "status",
          "id",
          "licenseNumber",
          "createdAt"
        ]
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Role"
              }
            ]
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserStatus"
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "driverProfile": {
            "nullable": true,
            "description": "Chỉ có với tài xế",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/DriverProfileDto"
              }
            ]
          }
        },
        "required": [
          "role",
          "status",
          "id",
          "email",
          "phone",
          "fullName",
          "createdAt"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserDto"
          }
        },
        "required": [
          "accessToken",
          "user"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
    }
  }
}